+++ src/config.rs
@@ -1,9 +1,11 @@
 // Copyright (C) 2023-2026 Daniel Mueller <deso@posteo.net>
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-#![cfg_attr(target_arch = "wasm32", expect(unused_imports))]
+#![cfg_attr(target_arch = "wasm32", expect(dead_code, unused_imports))]
 
+use std::fs::create_dir_all;
 use std::fs::read_to_string;
+use std::fs::write;
 use std::io::ErrorKind;
 use std::path::Path;
 use std::path::PathBuf;
@@ -68,6 +70,65 @@ fn load_config(path: &Path) -> Result<Config> {
 }
 
 
+fn save_config(config: &Config, path: &Path) -> Result<()> {
+  macro_rules! update {
+    ($doc:expr, $field:expr, as int) => {{
+      $doc[stringify!($field)] = toml_edit::value(i64::from($field));
+    }};
+    ($doc:expr, $field:expr) => {{
+      $doc[stringify!($field)] = toml_edit::value($field);
+    }};
+  }
+
+
+  let mut doc = load_config_doc(path)?.unwrap_or_default();
+  let current_result = from_toml_doc(doc.clone());
+
+  if let Some(dir) = path.parent() {
+    let () = create_dir_all(dir)
+      .with_context(|| format!("failed to create directory `{}`", dir.display()))?;
+  }
+
+  // We only want to write out the config if any of the values changed.
+  if current_result.as_ref() != Ok(config) {
+    let Config {
+      keyboard:
+        keys::Config {
+          auto_repeat_timeout_ms,
+          auto_repeat_interval_ms,
+        },
+      game:
+        game::Config {
+          start_level,
+          lines_for_level,
+          field_width,
+          field_height,
+          preview_stone_count,
+          enable_ai,
+          enable_dark_mode,
+        },
+    } = config.clone();
+
+    let keyboard = &mut doc["keyboard"];
+    update!(keyboard, auto_repeat_timeout_ms, as int);
+    update!(keyboard, auto_repeat_interval_ms, as int);
+
+    let game = &mut doc["game"];
+    update!(game, start_level, as int);
+    update!(game, lines_for_level, as int);
+    update!(game, field_width, as int);
+    update!(game, field_height, as int);
+    update!(game, preview_stone_count, as int);
+    update!(game, enable_ai);
+    update!(game, enable_dark_mode);
+
+    let () = write(path, doc.to_string())?;
+  }
+
+  Ok(())
+}
+
+
 /// A type representing the configuration of the program.
 #[derive(Clone, Default, Debug, PartialEq, Deserialize, Serialize)]
 #[non_exhaustive]
@@ -80,13 +141,20 @@ pub struct Config {
   pub game: game::Config,
 }
 
+#[cfg(not(target_arch = "wasm32"))]
 impl Config {
   /// Load the configuration from its default path on the file system.
-  #[cfg(not(target_arch = "wasm32"))]
   pub fn load() -> Result<Self> {
     let path = default_config_path().context("failed to retrieve program config directory path")?;
     let config = load_config(&path)?;
     Ok(config)
+  }
+
+  /// Save the configuration to its default path on the file system.
+  pub fn save(&self) -> Result<()> {
+    let path = default_config_path().context("failed to retrieve program config directory path")?;
+    let () = save_config(self, &path)?;
+    Ok(())
   }
 }
 
@@ -94,6 +162,12 @@ impl Config {
 #[cfg(test)]
 mod tests {
   use super::*;
+
+  use std::fs;
+  use std::thread::sleep;
+  use std::time::Duration;
+
+  use tempfile::NamedTempFile;
 
   use toml_edit::de::from_str as from_toml_str;
   use toml_edit::ser::to_string_pretty as to_toml_string;
@@ -128,5 +202,52 @@ start_level = 1
 lines_for_level = 10
     "#;
     assert!(from_toml_str::<Config>(config).is_ok());
+  }
+
+  /// Check that saving a [`Config`] does nothing if its fields didn't
+  /// change.
+  #[test]
+  fn save_config_no_change() {
+    let file = NamedTempFile::new().unwrap();
+    let path = file.path();
+
+    let config = Config::default();
+    let () = save_config(&config, path).unwrap();
+
+    let before = fs::metadata(path).unwrap().modified().unwrap();
+    // Sleep a tiny bit, because file systems may use coarse grained
+    // time stamps.
+    let () = sleep(Duration::from_millis(100));
+
+    let loaded = load_config(path).unwrap();
+    let () = save_config(&loaded, path).unwrap();
+
+    let after = fs::metadata(path).unwrap().modified().unwrap();
+    assert_eq!(before, after);
+  }
+
+  /// Verify that we write out a [`Config`] if it changed.
+  #[test]
+  fn save_config_with_change() {
+    let file = NamedTempFile::new().unwrap();
+    let path = file.path();
+
+    let mut config = Config::default();
+    let () = save_config(&config, path).unwrap();
+
+    let before = fs::metadata(path).unwrap().modified().unwrap();
+    let () = sleep(Duration::from_millis(100));
+
+    let loaded = load_config(path).unwrap();
+    assert_eq!(loaded, config);
+
+    config.game.enable_ai = !config.game.enable_ai;
+    let () = save_config(&config, path).unwrap();
+
+    let after = fs::metadata(path).unwrap().modified().unwrap();
+    assert!(after > before);
+
+    let loaded = load_config(path).unwrap();
+    assert_eq!(loaded, config);
   }
 }